Split debug logger into a module of it's own
