+++ packages/utils/telemetry-utils/src/index.ts
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-
+export * from "./debugLogger";
 export * from "./eventEmitterWithErrorHandling";
 export * from "./events";
 export * from "./logger";
